Move isHeterogeneous / isHomogeneous to DataStructures
