+++ Sources/DataStructures/CollectionExtensions.swift
@@ -1,0 +1,27 @@
+//
+//  CollectionExtensions.swift
+//  DataStructures
+//
+//  Created by James Bean on 6/29/18.
+//
+
+import Destructure
+
+extension Collection where Element: Equatable {
+
+    /// - Returns: `true` if there are one or fewer elements in `self`, or if all elements in
+    /// `self` are logically equivalent.
+    public var isHomogeneous: Bool {
+        guard let (head,tail) = destructured else { return true }
+        for element in tail where element != head { return false }
+        return true
+    }
+
+    /// - Returns: `false` if there are one or fewer elements in `self`, or if any elements in
+    /// `self` are not logically equivalent.
+    public var isHeterogeneous: Bool {
+        guard let (head,tail) = destructured else { return false }
+        for element in tail where element == head { return false }
+        return false
+    }
+}
