+++ Sources/Predicates/Predicates.swift
@@ -31,25 +31,6 @@ extension Sequence {
     }
 }
 
-extension Collection where Element: Equatable {
-
-    /// - Returns: `true` if there are one or fewer elements in `self`, or if all elements in
-    /// `self` are logically equivalent.
-    public var isHomogeneous: Bool {
-        guard let (head,tail) = destructured else { return true }
-        for element in tail where element != head { return false }
-        return true
-    }
-
-    /// - Returns: `false` if there are one or fewer elements in `self`, or if any elements in
-    /// `self` are not logically equivalent.
-    public var isHeterogeneous: Bool {
-        guard let (head,tail) = destructured else { return false }
-        for element in tail where element == head { return false }
-        return false
-    }
-}
-
 // FIXME: Move to different module than Predicates
 extension Sequence {
